requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Bootstrap-Flask==2.3.2
email-validator==2.1.0
pymongo==4.6.0
markdown2==2.4.12
bleach==6.1.0
faker==19.6.2
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
Flask-Principal==0.4.0